+++ src/main/java/racingcar/Application.java
@@ -1,6 +1,7 @@
 package racingcar;
 
 import java.util.Scanner;
+import racingcar.domain.CarRacing;
 
 public class Application {
     public static void main(String[] args) {
