+++ src/main/java/racingcar/domain/Car.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class Car {
     private final String name;
@@ -12,7 +12,9 @@ public class Car {
         return name;
     }
 
-    public void goForward() { position++; }
+    public void goForward() {
+        position++;
+    }
 
     public int getPosition() {
         return position;
