+++ src/main/java/racingcar/input/utils/racingtrytime/RacingTryTimeInputFromUserUtils.java
@@ -1,7 +1,7 @@
-package racingcar;
+package racingcar.input.utils.racingtrytime;
 
 import java.util.Scanner;
-import racingcar.input.InputPrinter;
+import racingcar.input.printer.InputPrinter;
 import racingcar.input.utils.racingtrytime.validators.RacingTryTimeValidatorUtils;
 
 public class RacingTryTimeInputFromUserUtils {
