+++ src/main/java/racingcar/utils/CarCreatorUtils.java
@@ -1,7 +1,8 @@
-package racingcar;
+package racingcar.utils;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import racingcar.domain.Car;
 
 public class CarCreatorUtils {
