+++ src/main/java/racingcar/utils/GoForwardOrStopUtils.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.utils;
 
 public class GoForwardOrStopUtils {
     public static final int MIN_GO_FORWARD_VALUE = 4;
