+++ src/main/java/racingcar/utils/randomnumber/OneRandomNumberGeneratorUtils.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.utils.randomnumber;
 
 import java.util.Random;
