+++ src/test/java/calculator/StringCalculatorTest.java
@@ -1,9 +1,9 @@
 package calculator;
 
-import org.junit.jupiter.api.Test;
-
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+import org.junit.jupiter.api.Test;
 
 public class StringCalculatorTest {
     @Test
