+++ src/test/java/racingcar/domain/CarRepositoryTest.java
@@ -1,13 +1,14 @@
-package racingcar;
+package racingcar.domain;
 
-import static org.assertj.core.api.Assertions.*;
-
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.ArrayList;
 import java.util.List;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.domain.CarRepository;
 
 public class CarRepositoryTest {
