+++ src/test/java/racingcar/input/utils/carname/CarNamesSplitterUtilsTest.java
@@ -1,6 +1,6 @@
 package racingcar.input.utils.carname;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
