+++ src/test/java/racingcar/input/utils/carname/validators/CarNameFormatValidatorUtilsTest.java
@@ -1,13 +1,14 @@
-package racingcar.input.utils.carname;
+package racingcar.input.utils.carname.validators;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThatCode;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.input.utils.carname.validators.exceptions.InvalidCarNameFormatException;
 import racingcar.input.utils.carname.validators.CarNameFormatValidatorUtils;
+import racingcar.input.utils.carname.validators.exceptions.InvalidCarNameFormatException;
 
 public class CarNameFormatValidatorUtilsTest {
 
@@ -31,7 +32,8 @@ public class CarNameFormatValidatorUtilsTest {
         inputCarNames.add("bbb111");
         inputCarNames.add("가나다1");
         inputCarNames.add("aaa1");
-        assertThatThrownBy(() -> CarNameFormatValidatorUtils.validateAllCarNamesFormat(inputCarNames))
+        assertThatThrownBy(
+            () -> CarNameFormatValidatorUtils.validateAllCarNamesFormat(inputCarNames))
             .isInstanceOf(InvalidCarNameFormatException.class);
     }
 
@@ -42,7 +44,8 @@ public class CarNameFormatValidatorUtilsTest {
         inputCarNames.add("aaa");
         inputCarNames.add("");
         inputCarNames.add("ccc");
-        assertThatThrownBy(() -> CarNameFormatValidatorUtils.validateAllCarNamesFormat(inputCarNames))
+        assertThatThrownBy(
+            () -> CarNameFormatValidatorUtils.validateAllCarNamesFormat(inputCarNames))
             .isInstanceOf(InvalidCarNameFormatException.class);
     }
 
@@ -53,7 +56,8 @@ public class CarNameFormatValidatorUtilsTest {
         inputCarNames.add("aaa");
         inputCarNames.add("b b");
         inputCarNames.add("ccc");
-        assertThatThrownBy(() -> CarNameFormatValidatorUtils.validateAllCarNamesFormat(inputCarNames))
+        assertThatThrownBy(
+            () -> CarNameFormatValidatorUtils.validateAllCarNamesFormat(inputCarNames))
             .isInstanceOf(InvalidCarNameFormatException.class);
     }
 }
