+++ src/test/java/racingcar/input/utils/racingtrytime/RacingTryTimeValidatorUtilsTest.java
@@ -1,13 +1,14 @@
 package racingcar.input.utils.racingtrytime;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThatCode;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.input.utils.racingtrytime.validators.exceptions.NotNaturalNumberRacingTimeException;
 import racingcar.input.utils.racingtrytime.validators.RacingTryTimeValidatorUtils;
+import racingcar.input.utils.racingtrytime.validators.exceptions.NotNaturalNumberRacingTimeException;
 
 public class RacingTryTimeValidatorUtilsTest {
