+++ src/test/java/racingcar/utils/CarCreatorUtilsTest.java
@@ -1,11 +1,13 @@
-package racingcar;
+package racingcar.utils;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Car;
+import racingcar.utils.CarCreatorUtils;
 
 public class CarCreatorUtilsTest {
