+++ src/test/java/racingcar/utils/randomnumber/OneRandomNumberGeneratorUtilsTest.java
@@ -1,11 +1,12 @@
-package racingcar;
+package racingcar.utils.randomnumber;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.utils.randomnumber.OneRandomNumberGeneratorUtils;
 
 public class OneRandomNumberGeneratorUtilsTest {
