+++ src/test/java/racingcar/utils/randomnumber/RandomNumbersGeneratorUtilsTest.java
@@ -1,10 +1,11 @@
-package racingcar;
+package racingcar.utils.randomnumber;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.utils.randomnumber.RandomNumbersGeneratorUtils;
 
 public class RandomNumbersGeneratorUtilsTest {
